Fix comparison between enums in LEX_UNTOK macro
This commit fixes the following compiler warnings from gcc 4.9.2 on
Debian 8:

> gcc -DHAVE_CONFIG_H -I.  -g -Wall  -c client-sock.c
> In file included from common.h:29:0,
>                  from client.h:29,
>                  from client-sock.c:38:
> client-sock.c: In function 'send_greeting':
> lex.h:89:14: warning: comparison between 'enum proto_toks' and
>       'enum common_tokens' [-Wenum-compare]
>      ( ((tok) < LEX_TOK_OFFSET) ? (tok) : ((tok) - LEX_TOK_OFFSET) )
>               ^

and

> gcc -DHAVE_CONFIG_H -I.  -g -Wall  -c server-sock.c
> In file included from common.h:29:0,
>                  from server-sock.c:43:
> server-sock.c: In function 'send_rsp':
> lex.h:89:14: warning: comparison between 'enum proto_toks' and
>       'enum common_tokens' [-Wenum-compare]
>      ( ((tok) < LEX_TOK_OFFSET) ? (tok) : ((tok) - LEX_TOK_OFFSET) )
>               ^

The common_tokens enum is a reserved range (>255) common to all
lexer enums here (of which there is only one, the proto_toks enum).
The compiler warning is silenced by casting the operands to ints.

+++ lex.h
@@ -86,7 +86,7 @@ typedef struct lexer_state *Lex;
 \************/
 
 #define LEX_UNTOK(tok) \
-    ( ((tok) < LEX_TOK_OFFSET) ? (tok) : ((tok) - LEX_TOK_OFFSET) )
+    ( ((int) (tok) < (int) LEX_TOK_OFFSET) ? (tok) : ((tok) - LEX_TOK_OFFSET) )
 /*
  *  LEX_TOK_OFFSET specifies the next available enumeration at which
  *    the array of strings supplied to lex_create (toks) can begin.
